-- added cpage module support ---- added contents, annots, changes, dipslay, fonts modules -- added DONOTUSE attribute -- changed DEPRECATED to DONOTUSE in some functions -- added listitem and improved double linked list implementation -- removed some vc2005 warnings -- added Page constants to pdfspecification -- splitted pdfoperators to two files (pdfoperatorsbase and pdfoperators) ---- removed unused funtions

+++ src/kernel/cstream.cc
@@ -223,7 +223,7 @@ CStream::setLength (size_t len)
 	//Lock
 	dictionary.lockChange ();
 
-	CInt _len (len);
+	CInt _len (static_cast<const int>(len));
 	dictionary.setProperty ("Length", _len);
 
 	// Unlock
